+++ src/UnderGUI/Inner/Drawer.py
@@ -1,0 +1,10 @@
+from UnderGUI.Utility import *
+
+# Draws to window client area. All methods are for override.
+class Drawer:
+    def __init__(self):
+        pass
+    
+    # Fills window client area with color.
+    def fill_view(self, color):
+        pass